backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# WebSocket and real-time features
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Background tasks
celery==5.3.4
kombu==5.3.4
billiard==4.2.0

# Image handling
Pillow==10.1.0
django-imagekit==4.1.0

# Authentication and security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# API documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
markdown==3.5.1

# Development and debugging
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# File handling
django-storages==1.14.2
boto3==1.29.7

# Pagination (built into DRF)

# Validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Notifications (using custom app)

# Cache
django-cachalot==2.6.1

# Admin improvements
django-admin-interface==0.26.0
django-colorfield==0.10.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
